Rename fast_beam.py to beam_search.py
